ファイルの整頓とViewの分割
